Added feature to parse in claibration frames separately. Remove image clipping from image alignment. Remove FWHM from header of reduced files. Created a script called reduction_functions.py to add the image reduction functions used in image_reduction.py
